Address Copilot's comments

+++ nasType/NAS_PartiallyAllowedNSSAI.go
@@ -3,11 +3,11 @@
 
 package nasType
 
-// PartiallyAllowedNSSAI 9.11.3.37
+// PartiallyAllowedNSSAI 9.11.3.103
 // SNSSAIValue Row, sBit, len = [0, 0], 0 , INF
 type PartiallyAllowedNSSAI struct {
 	Iei    uint8
-	Len    uint8
+	Len    uint16
 	Buffer []uint8
 }
 
@@ -17,10 +17,10 @@ func NewPartiallyAllowedNSSAI(iei uint8) (x *PartiallyAllowedNSSAI) {
 	return x
 }
 
-func (a *PartiallyAllowedNSSAI) GetIei() (iei uint8) { return a.Iei }
-func (a *PartiallyAllowedNSSAI) SetIei(iei uint8)    { a.Iei = iei }
-func (a *PartiallyAllowedNSSAI) GetLen() (len uint8) { return a.Len }
-func (a *PartiallyAllowedNSSAI) SetLen(len uint8) {
+func (a *PartiallyAllowedNSSAI) GetIei() (iei uint8)  { return a.Iei }
+func (a *PartiallyAllowedNSSAI) SetIei(iei uint8)     { a.Iei = iei }
+func (a *PartiallyAllowedNSSAI) GetLen() (len uint16) { return a.Len }
+func (a *PartiallyAllowedNSSAI) SetLen(len uint16) {
 	a.Len = len
 	a.Buffer = make([]uint8, a.Len)
 }
